requirements.txt
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
